feat ： 參數改成int

+++ Assets/Script/MainGame/Dice/Dice.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Dice : MonoBehaviour
 {
     float diceNum = 0, who;
-    public static float P1_totalNum = 0, P2_totalNum = 0, P3_totalNum = 0, P4_totalNum = 0, round;
+    public static int P1_totalNum = 0, P2_totalNum = 0, P3_totalNum = 0, P4_totalNum = 0, round;
     public static bool isThrow;
 
     public int min = 1;
